refactor InvSearchModule and InvUnloadModule to eliminate InvUnloadModule instance usage and centralize functionality in Utils
